python_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
pytz
numpy==1.26.4
notebook==7.1.3
pandas==2.2.2
scikit-learn==1.4.2
SQLAlchemy==2.0.30
mysql-connector-python==8.3.0
matplotlib==3.9.0